# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/items/get-all-items": {
      "get": {
        "description": "Public — no authentication. Returns the 20 most recently created items with the given tag, newest first, excluding expired and discontinued items. Out-of-stock items are still listed.",
        "operationId": "ItemsController_getAllItems",
        "parameters": [
          {
            "name": "tag",
            "required": true,
            "in": "query",
            "description": "Which tag to filter by.",
            "schema": {
              "enum": [
                "men",
                "women",
                "sale",
                "new"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponseDTO"
                }
              }
            }
          }
        },
        "summary": "List the 20 newest items carrying a tag",
        "tags": [
          "catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "Clothing Store API",
    "description": "Authentication, catalogue and ordering endpoints. Generated from the controllers, so it cannot drift from the code.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "httpOnly refresh cookie, scoped to /auth. Set at sign-in and rotated by POST /auth/refresh."
      }
    },
    "schemas": {
...
      "ItemListEntryDTO": {
        "type": "object",
        "properties": {
          "itemid": {
            "type": "number",
            "example": 12
          },
          "itemcode": {
            "type": "string",
            "example": "TEE-001"
          },
          "title": {
            "type": "string",
            "example": "Basic Tee"
          },
          "primary_pic_link": {
            "type": "object",
            "example": "https://cdn.example.com/tee-001-front.jpg",
            "nullable": true,
            "description": "The image tagged `primary`, or null if the item has none."
          },
          "secondary_pic_link": {
            "type": "object",
            "example": "https://cdn.example.com/tee-001-back.jpg",
            "nullable": true,
            "description": "The image tagged `secondary`, or null if the item has none."
          },
          "price": {
            "type": "number",
            "example": 2500
          },
          "discount_amount": {
            "type": "number",
            "example": 375,
            "description": "The largest currently-valid discount for this item, in currency. Percentage and fixed discounts are compared after conversion, so this is the amount actually taken off. 0 when nothing applies."
          },
          "message": {
            "type": "object",
            "example": "Only 2 left in stock!",
            "nullable": true,
            "description": "Set only when total stock across every variant is 1, 2 or 3. Null otherwise — including when the item is out of stock."
          }
        },
        "required": [
          "itemid",
          "itemcode",
          "title",
          "primary_pic_link",
          "secondary_pic_link",
          "price",
          "discount_amount",
          "message"
        ]
      },
      "ItemListResponseDTO": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemListEntryDTO"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
    }
  }
}